Extract status query runner partial
